Moved `ThemeProvider` functionality to `EnumProviders`
