+++ lib/providers/enum_providers.dart
@@ -1,4 +1,7 @@
+import 'package:dynamic_color/dynamic_color.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:lightmeter/data/models/dynamic_colors_state.dart';
 import 'package:lightmeter/data/models/ev_source_type.dart';
 import 'package:lightmeter/data/models/supported_locale.dart';
 import 'package:lightmeter/data/models/theme_type.dart';
@@ -8,8 +11,11 @@ import 'package:lightmeter/providers/service_providers.dart';
 import 'package:m3_lightmeter_resources/m3_lightmeter_resources.dart';
 
 enum _ListenableAspect {
+  brightness,
+  dynamicColorState,
   evSourceType,
   locale,
+  primaryColor,
   stopType,
   themeType,
 }
@@ -21,6 +27,22 @@ class EnumProviders extends StatefulWidget {
 
   static _EnumProvidersState of(BuildContext context) {
     return context.findAncestorStateOfType<_EnumProvidersState>()!;
+  }
+
+  static Brightness brightnessOf(BuildContext context) {
+    return InheritedModel.inheritFrom<_EnumProvidersModel>(
+      context,
+      aspect: _ListenableAspect.brightness,
+    )!
+        .brightness;
+  }
+
+  static DynamicColorState dynamicColorStateOf(BuildContext context) {
+    return InheritedModel.inheritFrom<_EnumProvidersModel>(
+      context,
+      aspect: _ListenableAspect.dynamicColorState,
+    )!
+        .dynamicColorState;
   }
 
   static EvSourceType evSourceTypeOf(BuildContext context) {
@@ -37,6 +59,14 @@ class EnumProviders extends StatefulWidget {
       aspect: _ListenableAspect.locale,
     )!
         .locale;
+  }
+
+  static Color primaryColorOf(BuildContext context) {
+    return InheritedModel.inheritFrom<_EnumProvidersModel>(
+      context,
+      aspect: _ListenableAspect.primaryColor,
+    )!
+        .primaryColor;
   }
 
   static StopType stopTypeOf(BuildContext context) {
@@ -59,11 +89,13 @@ class EnumProviders extends StatefulWidget {
   State<EnumProviders> createState() => _EnumProvidersState();
 }
 
-class _EnumProvidersState extends State<EnumProviders> {
+class _EnumProvidersState extends State<EnumProviders> with WidgetsBindingObserver {
   UserPreferencesService get userPreferencesService =>
       ServiceProviders.userPreferencesServiceOf(context);
 
+  late bool dynamicColor = userPreferencesService.dynamicColor;
   late EvSourceType evSourceType;
+  late Color primaryColor = userPreferencesService.primaryColor;
   late StopType stopType = userPreferencesService.stopType;
   late SupportedLocale locale = userPreferencesService.locale;
   late ThemeType themeType = userPreferencesService.themeType;
@@ -76,17 +108,59 @@ class _EnumProvidersState extends State<EnumProviders> {
             !ServiceProviders.environmentOf(context).hasLightSensor
         ? EvSourceType.camera
         : evSourceType;
+    WidgetsBinding.instance.addObserver(this);
+  }
+
+  @override
+  void didChangePlatformBrightness() {
+    super.didChangePlatformBrightness();
+    setState(() {});
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return _EnumProvidersModel(
-      evSourceType: evSourceType,
-      locale: locale,
-      stopType: stopType,
-      themeType: themeType,
-      child: widget.child,
+    return DynamicColorBuilder(
+      builder: (lightDynamic, darkDynamic) {
+        late final DynamicColorState state;
+        late final Color? dynamicPrimaryColor;
+        if (lightDynamic != null && darkDynamic != null) {
+          if (dynamicColor) {
+            dynamicPrimaryColor =
+                (_themeBrightness == Brightness.light ? lightDynamic : darkDynamic).primary;
+            state = DynamicColorState.enabled;
+          } else {
+            dynamicPrimaryColor = null;
+            state = DynamicColorState.disabled;
+          }
+        } else {
+          dynamicPrimaryColor = null;
+          state = DynamicColorState.unavailable;
+        }
+        return _EnumProvidersModel(
+          brightness: _themeBrightness,
+          dynamicColorState: state,
+          evSourceType: evSourceType,
+          locale: locale,
+          primaryColor: dynamicPrimaryColor ?? primaryColor,
+          stopType: stopType,
+          themeType: themeType,
+          child: widget.child,
+        );
+      },
     );
+  }
+
+  void enableDynamicColor(bool enable) {
+    setState(() {
+      dynamicColor = enable;
+    });
+    ServiceProviders.userPreferencesServiceOf(context).dynamicColor = enable;
   }
 
   void toggleEvSourceType() {
@@ -113,6 +187,13 @@ class _EnumProvidersState extends State<EnumProviders> {
     });
   }
 
+  void setPrimaryColor(Color primaryColor) {
+    setState(() {
+      this.primaryColor = primaryColor;
+    });
+    ServiceProviders.userPreferencesServiceOf(context).primaryColor = primaryColor;
+  }
+
   void setStopType(StopType stopType) {
     setState(() {
       this.stopType = stopType;
@@ -126,17 +207,34 @@ class _EnumProvidersState extends State<EnumProviders> {
     });
     ServiceProviders.userPreferencesServiceOf(context).themeType = themeType;
   }
+
+  Brightness get _themeBrightness {
+    switch (themeType) {
+      case ThemeType.light:
+        return Brightness.light;
+      case ThemeType.dark:
+        return Brightness.dark;
+      case ThemeType.systemDefault:
+        return SchedulerBinding.instance.platformDispatcher.platformBrightness;
+    }
+  }
 }
 
 class _EnumProvidersModel extends InheritedModel<_ListenableAspect> {
+  final Brightness brightness;
+  final DynamicColorState dynamicColorState;
   final EvSourceType evSourceType;
   final SupportedLocale locale;
+  final Color primaryColor;
   final StopType stopType;
   final ThemeType themeType;
 
   const _EnumProvidersModel({
+    required this.brightness,
+    required this.dynamicColorState,
     required this.evSourceType,
     required this.locale,
+    required this.primaryColor,
     required this.stopType,
     required this.themeType,
     required super.child,
@@ -144,8 +242,11 @@ class _EnumProvidersModel extends InheritedModel<_ListenableAspect> {
 
   @override
   bool updateShouldNotify(_EnumProvidersModel oldWidget) {
-    return evSourceType != oldWidget.evSourceType ||
+    return brightness != oldWidget.brightness ||
+        dynamicColorState != oldWidget.dynamicColorState ||
+        evSourceType != oldWidget.evSourceType ||
         locale != oldWidget.locale ||
+        primaryColor != oldWidget.primaryColor ||
         stopType != oldWidget.stopType ||
         themeType != oldWidget.themeType;
   }
@@ -155,9 +256,15 @@ class _EnumProvidersModel extends InheritedModel<_ListenableAspect> {
     _EnumProvidersModel oldWidget,
     Set<_ListenableAspect> dependencies,
   ) {
-    return (evSourceType != oldWidget.evSourceType &&
+    return (brightness != oldWidget.brightness &&
+            dependencies.contains(_ListenableAspect.brightness)) ||
+        (dynamicColorState != oldWidget.dynamicColorState &&
+            dependencies.contains(_ListenableAspect.dynamicColorState)) ||
+        (evSourceType != oldWidget.evSourceType &&
             dependencies.contains(_ListenableAspect.evSourceType)) ||
         (locale != oldWidget.locale && dependencies.contains(_ListenableAspect.locale)) ||
+        (primaryColor != oldWidget.primaryColor &&
+            dependencies.contains(_ListenableAspect.primaryColor)) ||
         (stopType != oldWidget.stopType && dependencies.contains(_ListenableAspect.stopType)) ||
         (themeType != oldWidget.themeType && dependencies.contains(_ListenableAspect.themeType));
   }
