+++ lib/screens/settings/components/theme/components/dynamic_color/widget_list_tile_dynamic_color.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:lightmeter/data/models/dynamic_colors_state.dart';
 import 'package:lightmeter/generated/l10n.dart';
-import 'package:lightmeter/providers/theme_provider.dart';
+import 'package:lightmeter/providers/enum_providers.dart';
 import 'package:lightmeter/res/dimens.dart';
-import 'package:lightmeter/utils/inherited_generics.dart';
 
 class DynamicColorListTile extends StatelessWidget {
   const DynamicColorListTile({super.key});
@@ -13,8 +12,8 @@ class DynamicColorListTile extends StatelessWidget {
     return SwitchListTile(
       secondary: const Icon(Icons.colorize),
       title: Text(S.of(context).dynamicColor),
-      value: context.listen<DynamicColorState>() == DynamicColorState.enabled,
-      onChanged: ThemeProvider.of(context).enableDynamicColor,
+      value: EnumProviders.dynamicColorStateOf(context) == DynamicColorState.enabled,
+      onChanged: EnumProviders.of(context).enableDynamicColor,
       contentPadding: const EdgeInsets.symmetric(horizontal: Dimens.paddingM),
     );
   }
