+++ lib/screens/settings/components/theme/components/primary_color/widget_list_tile_primary_color.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:lightmeter/data/models/dynamic_colors_state.dart';
 import 'package:lightmeter/generated/l10n.dart';
-import 'package:lightmeter/providers/theme_provider.dart';
+import 'package:lightmeter/providers/enum_providers.dart';
 import 'package:lightmeter/res/dimens.dart';
 import 'package:lightmeter/screens/settings/components/theme/components/primary_color/components/primary_color_picker_dialog/widget_dialog_picker_primary_color.dart';
-import 'package:lightmeter/utils/inherited_generics.dart';
 
 class PrimaryColorListTile extends StatelessWidget {
   const PrimaryColorListTile({super.key});
 
   @override
   Widget build(BuildContext context) {
-    if (context.listen<DynamicColorState>() == DynamicColorState.enabled) {
+    if (EnumProviders.dynamicColorStateOf(context) == DynamicColorState.enabled) {
       return Opacity(
         opacity: Dimens.disabledOpacity,
         child: IgnorePointer(
@@ -31,7 +30,7 @@ class PrimaryColorListTile extends StatelessWidget {
           builder: (_) => const PrimaryColorDialogPicker(),
         ).then((value) {
           if (value != null) {
-            ThemeProvider.of(context).setPrimaryColor(value);
+            EnumProviders.of(context).setPrimaryColor(value);
           }
         });
       },
